Move FieldCollection et al. back to Field
