+++ packages/kiwi-react/src/bricks/Switch.tsx
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 import cx from "classnames";
 import * as Ariakit from "@ariakit/react";
-import { useFieldDescribedBy } from "./Field.js";
-import { FieldControl, forwardRef, type FocusableProps } from "./~utils.js";
+import { FieldControl, useFieldDescribedBy } from "./Field.js";
+import { forwardRef, type FocusableProps } from "./~utils.js";
 
 type InputBaseProps = Omit<FocusableProps<"input">, "defaultValue" | "value">;
